Prevent loading templates out of the search paths
This check prevents someone to use a relative path to a template in a directory outside the search paths. For example: "../../something/bad.html".

I am using nunjucks in a web service where users can edit their own templates and of course I don't want to let them to find a system file or use a template of another user.