Add option to automatically indent blocks
Add an `indentBlocks` options that will automatically indent the block
content of extending templates to the level of the block in the base
template that originally defined the block for the first time.

This ensures that the final rendered output will follow the indentation
outlined in the original base template.

This functionality also works through multiple levels of inheritance and
for nested blocks that are first defined in an extending template.

Example base template:

    <div>
      {% block A %}{% endblock %}
      <div>
         {% block B %}
         default content
         {% endblock %}
      </div>
    </div>

Example extending template:

    {% extends 'base.html' %}
    {% block A %}
    <h1>Example content</h1>
    {% endblock %}

    {% block B %}
    This content will have the original indentation.
    {% endblock %}

Rendered result:

    <div>
      <h1>Example content</h1>
      <div>
        This content will have the original indentation.
      </div>
    </div>