Fix docs build.
Since nunjucks is documented using Jekyll, which uses Liquid templates, which
use the same basic syntax as Jinja2/nunjucks, we can't document an example of
{% raw %}...{% endraw %} without breaking the documentation template (which
wraps most of the docs inside {% raw %}).

This is pretty unfortunate; we really should not document nunjucks using a
template language with the same syntax.